\name{set.coord.type}
\alias{set.coord.type}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
set user's coordernate system
}
\description{
坐标类型，可选参数，默认为bd09ll。允许的值为：bd09ll（百度经纬度坐标）、bd09mc（百度摩卡托坐标）、gcj02（国测局加密坐标）、wgs84（gps设备获取的坐标）。
}

\examples{
set.coord.type(coord_type)
}

